al registrarte pilla el id del usuario y sube y busca servicios con este id y ademas elimina los servicios añadiendo el id en el input de misServicios

+++ src/main/java/com/example/demo/Servicio.java
@@ -19,7 +19,7 @@ public class Servicio {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int servicioid;
+	private long servicioid;
 	private String nombre;
 	private String descripcion;
 	private String categoria;
